JFactory: limit cache reallocation to caches in use
JFactory starts with no caches, and allocates caches on demand. Therefore,
caches that are ever used are kept and grown with the number of BDDs for all
time. However, most computations go through stages, and caches for specialized
operations become unused.

When resizing, free caches that have not been heavily used since the last
resize operation.