BroadcastL3Adjacencies: do not connect interfaces mentioned in L1 to hub
Previously, the code relied on L1 Logical Topology, but that is too clean: any
edge where an interface is not in the snapshot or is down is removed before we
see it. As a result, we cannot rely on the final logical topology to determine
whether a specific interface has an L1 edge.

Rewrite the code, mainly moving copies of the TopologyUtil code into
L3AdjacencyComputer.